+++ src/main/java/ConcurrentUtils/CountDownLatchExample.java
@@ -1,4 +1,4 @@
-package ConcurrentUtilsExample;
+package ConcurrentUtils;
 
 import java.util.concurrent.CountDownLatch;
