+++ src/main/java/ConcurrentUtils/CyclicBarrierExample2.java
@@ -1,4 +1,4 @@
-package ConcurrentUtilsExample;
+package ConcurrentUtils;
 
 import java.util.concurrent.BrokenBarrierException;
 import java.util.concurrent.CyclicBarrier;
