+++ src/main/java/ConcurrentUtils/SemaphoreExample.java
@@ -1,4 +1,4 @@
-package ConcurrentUtilsExample;
+package ConcurrentUtils;
 
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
